+++ src/acc_app_optimisation/job_control/rl/__init__.py
@@ -6,8 +6,7 @@
 
 from . import agents
 from .agents import ALL_AGENTS, AgentFactory, BaseAlgorithm
-from .execute import ExecJob, ExecJobBuilder
-from .train import CannotBuildJob, TrainJob, TrainJobBuilder
+from .execute import CannotBuildJob, ExecJob, ExecJobBuilder
 from .wrapper import PreRunMetadata
 
 __all__ = [
@@ -18,7 +17,5 @@ __all__ = [
     "ExecJob",
     "ExecJobBuilder",
     "PreRunMetadata",
-    "TrainJob",
-    "TrainJobBuilder",
     "agents",
 ]