+++ CHANGELOG.md
@@ -10,7 +10,9 @@ SPDX-License-Identifier: GPL-3.0-or-later OR EUPL-1.2+
 
 ## Unreleased
 
-No changes yet!
+### Visible changes
+
+- Remove the tab for training RL algorithms due to lack of use.
 
 ## v0.16.1